Update to exercise 1 and 2
In the homework exercise 1 on page: 64 and 65 of the book it says: "If the user enters anything other than a number, detect their mistake using try and except and print an error message and skip to the next number."  The previous code submitted as a solution did not fulfill the requirement stated in the book.  So I wrote my own code until I figured out how to make the program "detect their mistake using try and except and print an error message and skip to the next number" using the "continue" function on line 23.